to be upstreamed: Allow registering memory from current address space
bug: 146066070
bug: 153580313

This is good for enabling non-exportable Vulkan host coherent memory,
along with anything else where an exportable OS object isn't supported.

This CL introduces two structures:

1. ExternallyMappedHostMemory, which represents an existing buffer
(already allocated or mapped) in the current address space, for purposes
of sharing device memory to the guest in the case where the device
memory is not compatible with the mmap interface. This is common in
Vulkan when VkDeviceMemory is host visible but not external, or external
but based on an opaque fd.

We make the assumption that the actual host address range backing the
region outlives ExternallyMappedHostMemory.

The intended use of ExternallyMappedHostMemory is for it to be inherited
as the base of another Trait object, which will specify any additional
cleanup when it's dropped from the hypervisor's memory tables, and can
override how to obtain the host pointer and size.

2. Usually, we would send such memory requests over a socket to the main
thread. However, since these new objects require more metadata than
other requests that are sent over the wire (because there's information
about inheritance and refcounts), we also introduce
ExternallyMappedHostMemoryRequests, which communicates a single request
at a time to the main thread but can handle passing over first-class
Rust objects.

We then initialize a single instance on boot and pass an Arc<Mutex>
reference to it for users, starting with virtio gpu.

Note that this ExternallyMappedHostMemory will not work in the sandbox
case. In the sandbox case, we will then have to figure out how to
serialize/deserialize ExternallyMappedHostMemory requests over a socket.

Change-Id: I3b099b308aec45a313a8278ed6274f9dec66c30b