Fix sandboxing of virtio-block on aarch64
Seen when running with virtio-block enabled on rockpi4b:

[ERROR:src/linux.rs:1885] child virtio-pci (virtio-block)
  (pid 14298) died: signo 17, status 31, code 2

Looks like when gettid() was factored out to common for arm/x86 in
I129644273f2f02fe917255c7157c48b99c329045, aarch64 was left alone. Until
recently virtio-block wasn't using gettid, but now it seems to be, so we
need to make the same change for aarch64.

Change-Id: Ie78c127178563d3b16236db71b187555c6cc5b6e
Signed-off-by: Alistair Delva <adelva@google.com>