ANDROID: Add code to generate inlined seccomp policy
The upstream crosvm seccomp policy for multi-process jailing contains
hardcoded rereferences to /usr/share/policy/crosvm/common_device.policy,
which only works on ChromeOS. Using relative paths isn't supported by
minijail, so it's easier to just inline the policy includes at build
time.

This change adds a script to generate Android.bp files which themselves
use genrules to generate the seccomp policy files on the fly. There are
currently three sets of policy files:

 - Host Linux x86_64
   These are used by the host crosvm built from source for x86_64.

 - Host Linux aarch64
   These are used by a host crosvm prebuilt for aarch64.

 - Target (x86_64 or aarch64)
   These are used by the Android target build of crosvm.

The script also generates a "host_package.mk" file for virtual platforms
like Cuttlefish which need to enumerate files to install on the host.

The seccomp policy for x86_64 also uses @frequency, which suffers from a
similar limitation in minijail, but as this is just a performance
optimization, drop the lines and don't install the frequency hints.

Bug: 166118192
Change-Id: I98629f2162fa956c00926baf0ebe7b6abc6e31d8