Enable crosvm for linux_bionic_arm64
linux_bionic_arm64 is the new target for the cross-compiled Linux/ARM
host. crosvm (and its dependencies) are enabled for the new target.
Features are enabled for the target. The exact list of the feature was
retrieved from device/google/cuttlefish_vmm/rebuild-internal.sh
and Cargo.toml files that we used to build crosvm for the target out of
the Android build system.

In addition, the host.* properties are now changed to linux_glibc.*
because host.* covers both linux_glibc (non-cross target) and
linux_bionic (cross target). Some settings (e.g. "-C panic=abort") are
not applicable to linux_bionic targets.

Bug: 159685774
Test: HOST_CROSS_OS=linux_bionic m crosvm
Test: copy the contents under out/soong/host/linux_bionic-arm64 to
a QEMU-based Linux/ARM VM and run crosvm from it.

Change-Id: Icfd9a3d56a0632e0347eeeb168d1380c7f0165db