Removed code path specific to psci 1.0
The code being removed is broken. The compatible field, if populated
with multiple strings, needs to have each one separated with null
characters.

This code as is breaks fdt parsing for the psci node. Additionally, the
FDT entry for psci's compatible string does not need to include psci
0.2. The linux kernel will load the firmware for 0.2 as well as 1.0 psci if
1.0 is detected.

The simplest solution is to delete it all and include the appropriate
psci version number in the node.

Bug: 177926450
Change-Id: I2e0a519cbf00c833c67ffe9bfde1e804d7b62e01