ANDROID: crosvm: Run cargo2android after merge
Set "no-pkg-vers" in the hypervisor crate. cargo2android is trying to
get the pkg version for the src files in the "tests" directory and those
don't seem to have pkg versions. There is currently no way to disable it
for just part of the crate AFAIK.

Had to manually enable a feature in the vm_memory crate.

Test: TH
Bug: 239563396
Change-Id: Idf347a7597d714acc46e6f012b5701c9c0aa722e