virtio: video: encoder: Handle SET_PARAMS_EXT's resource_type.
Previously we kept using the initial value and while this was fine for
ARC because VirtioObject was the initial value negotiated by the kernel
and ARC only used VirtioObject anyway, but this broke when using FFmpeg
which would use the mmap buffers of V4L2.

Handle the type switch like we did in the decoder.

BUG=None
TEST=arc.VideoDecodeAccel*_vm,arc.VideoEncodeAccel*_vm on hatch

Change-Id: I6f1c2d1bdaf18efb559528dfadb10baf32dbb757
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3840991
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Tatsuyuki Ishi <ishitatsuyuki@google.com>
Commit-Queue: Tatsuyuki Ishi <ishitatsuyuki@google.com>