Revert "Merge remote-tracking branch 'aosp/upstream-main' into m..."
Revert "sepolicy: add new genfs_contexts for APCI device wakeup"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/4703e299-8288-473f-9701-3408c33cae1b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I27b95aab8fa549192aa03243740d797a7557ba17