Revert "Merge remote-tracking branch 'aosp/upstream-main' into m..."
Revert "sepolicy: add new genfs_contexts for APCI device wakeup"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/7eb4c6ff-4c17-4881-8285-2e6ca38efa4c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I74c71db58263b5e35e25f3ce93ee04a56624a84e