devices: gpu: Support display hotplug
Adds `crosvm gpu <>` commands which allows for interactively
adding/listing/removing virtio-gpu displays/scanouts.

Also updates the virtio gpu device to use a Map<u32, Scanout>
instead of a Vec<Scanout> and always report the maximum number
of displays (16) with enabled/disabled. This helps handle the
case of removing a display that is not the last display number.

Bug: b/163864461
Test: cargo build
Test: cargo build (with gpu feature disabled)
Test: Android build
Test: launch Cuttlefish
Test: crosvm gpu list-displays <socket>
Test: crosvm gpu add-displays
      --gpu-display=width=900,height=600
      --gpu-display=width=600,height=480
      <socket>
Test: crosvm gpu list-displays <socket>
Test: crosvm gpu remove-displays
      --display-id=1
      <socket>
Change-Id: I3c346f61c2b9718494e994e764450d9dc8b484f6