ANDROID: Add a month/day suffix to merge script branch
`repo start merge` will likely checkout the branch that was last used
for the crosvm merge that is out of date.
`merge` will also not work for topic names for the CL if we need to
include other CLs from other projects in the same topic.
So, add <month/day> suffix to the branch name.
ex: `merge_Sept27`

Bug: 249152023
Test: ./merge-setup.sh
Change-Id: I0609632466ce57c841661948adf13226b27d3125