Revert "Merge commit 'ba3e2fa85261f0fcb4faee865ebe59278f3d5204' ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5663fdf2-2cb3-4c43-9cbd-e0a2eee1d06a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1460fe661f333fedcd5993c99b1d7beb7b09543f