Revert "ANDROID: update import script to use cargo_embargo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/da368da7-6ce5-4cdb-8c86-d0b8f6f75571). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I89d778f423bbc6a1082fe71bf87f9a76ebfee14b