Revert "Merge remote-tracking branch 'aosp/upstream-main' into merge"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/217483a1-374d-4b87-8e4c-e9a52b7b6298). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I532d64bcae4c142480b956207323d0ddf58fa3e1