Revert "ANDROID: Run cargo_embargo after merge"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/217483a1-374d-4b87-8e4c-e9a52b7b6298). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1bcb754b77d5bdf104be2957d56e016368f8180c