[WIP] Support --host-cpu-topology on arm64
The --host-cpu-topology command line option can be used on x86_64 to
create a 1:1 mapping between physical CPUs and guest's vCPUs. This patch
allows the same option to be used with arm64 but with slightly different
semantics.

This is because testing on Android devices showed that a 1:1 mapping
does not yield very good performance. Given that Android devices
proactively suspend CPU cores, a 1:1 mapping prevents the host scheduler
from making use of currently online cores. This can be improved by
setting the affinity mask of a vCPU to the whole cluster, but the
approach taken in this patch is to set vCPU capacities 1:1 and affinity
mask to any core with capacity equal or greater than that of the given
vCPU. This gives the host scheduler the most freedom while guaranteeing
the vCPU capacity to the guest.

To this end, the guest also needs to be informed about the vCPU capacity
and clusters. Both of these options are ignored on x86_64 and now cannot
be provided together with --host-cpu-topology. The information on arm64
is obtained through sysfs and passed to the guest via DT.

Bug: 266664564
Change-Id: I8f6d5cea7d8861e27f02ec55b50c0d12e1c2866b