Revert "ANDROID: Disable flaky io_uring test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/118563de-1c67-4b8b-a7ae-bed9ec80ca23). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I97848cf58e9bcdc563de9f06c898ebd3065501f0