vm_memory: Add early support for restricted memory
kernel.org is preparing to introduce support for a new type of memfd:
restricted_memfd ([1], [2]). This new type of memfd is to be used for
guest-private memory which shouldn't be available to the host. New
ioctls and separate memory mappings are needed for this type of memory.

Although restricted_memfd support is not yet available in any kernel
crosvm supports, Gunyah requires separate ioctls to differentiate
between "protected" (restricted) memory and "unprotected" (shared)
memory, even though the memory is backed by a memfd in either case.

Introduce the plumbing to differentiate between the restricted memory
and shared memory for Gunyah. Instead of a restricted_memfd as the
BackingObject, the common SharedMemory backing object is used.

[1]: https://lore.kernel.org/all/20221202174417.1310826-1-tabba@google.com/
[2]: https://lore.kernel.org/all/20221202061347.1070246-1-chao.p.peng@linux.intel.com/

Bug: 232360323
Change-Id: Id26bd645cb857accd69bb2db128c4874ce5c3e6f
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>