Revert "Merge remote-tracking branch 'aosp/upstream-main'"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e21b89c3-f235-46b5-9c19-ed44d8f3aa21). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia47aa67975ddaf055333d50cb18724ef8035de6b