Revert "UPSTREAM: Fix building sock_ctrL_msg.rs against musl libc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/32a112bd-16ac-43f7-b558-b24680b273b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie789b16c7b1b2fbf1ae7625bf8512dc3d3e19969