Revert "ANDROID: run cargo_embargo after merge + fix seccomp script"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81b5fdf9-ec65-4451-86f9-76bc6b308114). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice24139578e3a189243ba57a7fc194807c4e3935