devices: gpu: build Rutabaga in on_device_sandboxed()
... to give rutabaga and the underlying render server a chance to fully
initialize before the guest OS starts. The Virtio GPU kernel module has
a 5 second timeout during init to wait on capset info from the host. The
Nvidia driver on fresh GCE instances has been observed to take ~2-3 seconds
for each of GL and VK initialization. This, combined with the Gpu not
starting host renderer initialization until activate(), often leads to the
host not responding in time and ultimately device failures on Cuttlefish.

Bug: b/287523712
Test: cvd start --gpu_mode=gfxstream
Change-Id: I400d12ad911b3e2875230ddcc06518c8562604eb