crosvm: Gunyah: Add support for Qualcomm specific Authenticated VMs
Qualcomm specific Authenticated VMs (called Auth VMs) have specific
requirements w.r.t memory and authentication mechanism to be used for
the provided VM image.

These VMs boot from physically contiguous cma/carveout memory due to
the underlying limitation of the firmware that authenticates the VM
image. These VMs will thus have to use the fixed memory size for each
memory slot. The memory layout/size will be reserved by the kernel as
per it's memory layout. The memory slots supported for these VMs would
be either one slot for VM main memory or two slots based on whether it
is configured with swiotlb or not.

The kernel exposes an ioctl on Gunyah VM fd to query the
available/reserved memory sizes for each slot. It is only after the
memory sizes are known, crosvm would prepare the guest memory region by
doing mmap() on each memory region and start the VM normally. The kernel
would take care of authenticating the loaded VM image by the firmware
and then proceeds to boot it normally.

To be able to differentiate this type of VM, the specific/reserved vmid
of the virtual machine would have to be passed on to the kernel as part
of GH_CREATE_VM ioctl.

To achieve above, the key changes introduced in this change are:

1. Add two new cmdline parameters to identity these VMs-
	--use-fixed-memory
	--vm-id <vmid>

2. Changes to VM start sequence in hypervisor/gunyah crate are-
	a. create VM instance using GH_CREATE_VM ioctl with the provided vm_id
	b. Query and get the memory slots size using GH_AUTH_VM_GET_MEM_PARAMS ioctl
	c. map the backing cma/careveout guest memory region into crosvm by
	   doing mmap()

TODO:
To autogenerate the bindings used in this patch

Change-Id: I4e0fbf82cb17433766ba807476c8fb2c3d0a7b0f
Signed-off-by: Sahitya Tummala <quic_stummala@quicinc.com>