Revert "Merge remote-tracking branch 'aosp/upstream-main'"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/171a8604-4c32-41bb-b7b2-db70c06daa68). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99e55ce7596c107934f7071e0f0994c570d609df