Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dca8c2b2-f68d-4417-9528-b0f99d0ecd94). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib7129db67251a3c1decb72238cd1d3d7c7312ac4