Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64e8a597-03df-4910-a4a2-ae0ee366904d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id4b9f1565d9abd2dcf4fc8152f983b6db981fdcd