Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/322c4738-6f07-424f-9166-02ff9ca5d12c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e783926759807c4c0ccc252e8f6ac159ddc58a8