Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50787364-94e6-49c3-80dc-5c0adc1ebee1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iea60ac3c75af1b654f9665631ecc7a485c75461c