Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c3d37578-d3b3-4e23-8530-7c272d940b47). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifcfb6883661cb05e60c3f9fcc41f667d77956652