Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13d4444c-5184-4b24-9141-af6b58c13265). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id582689c76b2449d6f3f25666b3cfb91d0dce2df