Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb349c73-bcab-4781-a1bb-18775bfce60f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I03b9d85f5b53280ebf409313e9b5bc88a6d44138