Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/222a7f02-eb21-4e01-ba64-3a235b45aaf3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic197fba1122a12823bd198f855a8dcccc489c959