Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8be36a56-a347-4ea3-aff9-1bcecf4d7428). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I332608e54a056aedabf2683c3793625d400992f3