Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8786490-9b8b-4cb9-81c9-1d29672888fc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib24d16208f5745253079a6fa6a392654643d3c81