Revert "ANDROID: remove obsolete visibility rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6a67c91-7b42-4999-84cc-a89b7cb0998e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id650bd77f1cf3c984f7d1ac391cea6579ca33579