hypervisor: x86: add get_clock_state for snapshot/restore
Get/Set Clock State when snapshotting/restoring the VM.
When a restore happens, the clock goes back in time. Provide a way to
sync the clock with the host.
This is achieved by saving the clock state, and setting it back on
restore.
NOTE: Only the flags of clock state are set on restore, not the clock
itself. This will however sync the clock with the host.

BUG=N/A
Test=boot vm, snapshot, restore

Change-Id: I7594eae015d8e9231d1a372e768f745a83a0409c