ANDROID: fix "metrics" diffs using new cargo_embargo feature
The "metrics" crate locally renames the "metrics_generic" crate to
"metrics_protuct" using a cargo feature. We can't replicate that in
Soong, so instead we change the crate name of the generated
"metrics_generic" module to "metrics_protuct". We can get away with that
because "metrics" is the only module depending on it.

Bug: 329312851
Test: ran cargo_embargo and built crosvm
Change-Id: I404007398a4b427c82518ee2db948f2afb8dfe71