hypervisor: gunyah: Mark shm devices as optional
When using demand paging, userspace shouldn't provide shm vdevice in the
VM configuration. The shm vdevice is used by RM to automatically accept
and map the mem parcel. With demand paging, the kernel wouldn't be
creating mem parcels and Gunyah complains about that no mem parcel was
provided for the shm vdevice. Gunyah allows us to add "optional" tag to
the vdevice and it will be ignored if no mem parcel is prvoided, as is
the case for demand paging.

We could create additional downstream UAPI to indicate whether demand
paging is being used and not create the vdevice node at all, but the
approach of adding "optional" property was preferred to avoid creating
downstream UAPI.

Bug: 330496811
Change-Id: Ib070e9d453af88a7af5f8c240f9a22c7b744c987
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>