Revert "Merge remote-tracking branch 'aosp/upstream-main' into merge"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/def8c6dc-6435-4929-9da7-f35fff2baf9b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iebfb9fbfd76e32138b2c53478275b213154b16a9