Refactor android display backend
This change refactors the C interface between crosvm and the android
display backend which is designed to be implemented outside of crosvm.

The below is a few important changes:

* The new interface allows crosvm to draw directly on Android surface.
  Previously, we allocated a framebuffer in crosvm, let crosvm draw on
it, and passed it to the Android-side backend where the framebuffer
is copied to Android surface using CPU.

* Respect stride set by Android. When a buffer is created on the Android
  side, the stride is usually set to a multiple of 16 for maximum
performance. This change reports that stride back to crosvm so that it
can respect the stride. Previously, we used a stride value that is equal
to the screen width which may not be multiple of 16.

* The error callback function pointer is passed to the Android side only
  when the context is created. The Android side is expected to keep the
pointer along with the context and use it when something goes wrong.

Bug: 331900473
Test: N/A
Change-Id: I3e5b05407c41413ea55bba37cb7f34f504799734