Revert "link gfxstream_backend and virglrender statically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb6ededd-5bbb-4136-93e8-3e82fbd32b6f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If28d33f75a2cbaaf9e6baf9b465e0b806fa7a10d