Revert "link gfxstream_backend and virglrender statically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd91dd50-17ba-435a-b3ce-eaf174e55238). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55499c0bf184dd483e7c65dd760afc119dcd0d17