Revert "UPSTREAM: aarch64: fdt: Fix CPU compatible to "arm,armv8""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/edfccbb4-d25c-4a7f-9173-c2c7404979f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0832926b9e7d9e18d8b57eb19c8a79186361256c