hypervisor: x86: add get_clock_state for snapshot/restore
Get/Set Clock State when snapshotting/restoring the VM.
When a restore happens, the clock goes back in time. Provide a way to
sync the clock with the host.
This is achieved by saving the clock state, and setting it back on
restore.
NOTE: Only the flags of clock state are set on restore, not the clock
itself. This will however sync the clock with the host.

Bug: 334994657
Test: boot vm, snapshot, restore

Change-Id: I7d894c79361edfef2cfadc5273e05509bce62601