Implement virtio-snd aaudio stream to support audio in android
note: this CL is uploaded to aosp for review. planned to upstream to
crosvm.
This change introduces feature audio_aaudio on top of the support of
virtio-snd-multiple-backend. The AndroidAudioStream uses the Android
Audio NDK library to play audio. A libaaudio_stub is provided for
building and testing on non-android environment.

Test: Play Youtube video on panther
Test: ./tools/dev_container
Test: cargo build -F audio_aaudio
Test: ./tools/run_tests
Test: ./tools/presubmit
Bug: 325930215
Change-Id: I7c275433bf94d4a347720ce0ffa9127ffcb66a1e