TEST-ONLY: rutabaga_gfx: gralloc: don't always start Vulkano
RutabagaGralloc is used for two things:

   1) Allocating swapchain buffers
   2) Using Vulkan to map buffers via hypervisor-specific syscalls

(2) is done in the main crosvm process, and could be sensitive
to performance considerations.  The "cuttlefish_fastboot_flashing_test"
seems to fail when Vulkano is initialized.  Maybe initializing Vulkan
takes too long, who knows.

Only Windows uses (2), and Linux doesn't really need it right now.
All Linux platforms could benefit from (1), so modify
RutabagaGralloc initialization so Vulkan is available for (1) but
not (2) on Linux.

BUG=336291294
TEST=compile

Change-Id: Ic753298230f05e4c314f65b3efbb362c76e21cd9