UPSTREAM: fix memory size when static swiotlb is enabled
Previously, if `--mem` was set to 100MB and `--swiotlb` to 1MB, then
we'd essentially allocate 101MB of RAM for the guest and tell it about a
single 101MB memory region in the FDT (adjacent regions are merged in
the FDT). In contrast, for hypervisors that don't require static
swiotlb, we'd only allocate 100MB of RAM (and then the guest would need
to choose 1MB of its own memory to dedicate to this purpose).

Now both cases will get the same RAM size.

Removed the swiotlb reference from memory_layout.md since it is part of
the RAM region from the perspective of the guest. The need for the
static swiotlb to be created as a separate memslot is more of a
hypervisor implementation detail I think.

BUG=b:346770542

Change-Id: I60b10521635c16fa77b4320de0e11f03ebeca0c5
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5637774
Reviewed-by: Elliot Berman <quic_eberman@quicinc.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Richard Zhang <rizhang@google.com>