Revert "Running cargo_embargo after the merge"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6fd0639-b739-45b2-9dd2-23b753af83a3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349617789
Change-Id: Ia5ee2a16e92ab9a27082410757f5b35c49f5c74e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com