Revert "Merge remote-tracking branch 'aosp/upstream-main' into merge"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db4499d1-052d-4566-8738-43bd176ca0e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349612065
Change-Id: I77acb16937b28bb866d2fad1bec47115ef477fa9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com