Revert "UPSTREAM: ext2: use os::unix rather than os::linux APIs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db4499d1-052d-4566-8738-43bd176ca0e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349612065
Change-Id: I5539250355405a4655aaef5c3fa40d1633dfff2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com