Revert "Fix riscv crosvm build error"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f8bbea8f-3a47-4f73-9dfd-794bf81f0910). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2159098477291d7f1f018a9e8986bb24192fd0f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com