Revert "Enable aaudio backend"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/675dbfd5-bf20-4732-a864-369731712e82). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I662a6055179c90231ecdcb9d648d3600afe4600b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com