Revert "UPSTREAM: Implement virtio-snd aaudio stream to support ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/675dbfd5-bf20-4732-a864-369731712e82). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I67ad0e654a5f6268a38ab1dcc660fe885b252710
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com