Revert "rutabaga_gfx: add Kumquat server and virtgpu_kumquat support"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/34bdff6d-d7f3-498a-b364-3e27d34b5e36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7ce43e64a4f8120973b0fb22fd0b41f852880a5e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com