devices: virtio_gpu: Keep state across sleep/wake
Prior to this change, the

  std::mem::drop(worker.state);

would take a snapshot and then completely destroy the underlying
renderer on sleep and then create a new renderer and attempt to
restore the previous state on wake. Sleep and waking without a
snapshot load should be a non destructive operation.

This change attempts to mainly update the worker to be persistent
and not be completely destroyed on sleep and wake.

Bug: b/369615058
Test: build
Change-Id: I6427c27ee4c21c3e7b1095fbfffb69afcd2a8d24