Revert "UPSTREAM: virtcpufreq_v2: Place vcpu worker threads dire..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/746b4dfd-3ae6-4f88-b4e7-3265fda544fa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I803b85ecddc440232a6ba0c8fa21e57e2f8a59ae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com