Revert "devices: serial: Support UnixStream for data source and ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1727d91d-46be-480c-abc8-5cb5210b6d7b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id5d0e1f9f02f08f52807be039801463f2b559c89
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com