Revert "input: Add custom input device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f4b96252-864a-4a5e-afd3-dbbb5fcc9a01). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icba8bc69be499f3743f0dec5663e33f91203ca4b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com