Revert "Prepare for thiserror 2.0."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c684163b-9dd9-4c45-8e9f-5780e5d6d54e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib07a4e5b5762b67ccd9aae84564be4428b554ef5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com