snapshot: hide serde_json::Value behind a newtype
Copy of http://crrev.com/c/6161187

The heavy use of `serde_json::Value` in the snapshot APIs makes it
difficult to experiment with other serialization frameworks. The vast
majority of the code just needs to supply a type that is serde
compatible and doesn't need to know what the format is.

On an aarch64 `--profile chromeos` build, the binary size barely changed
(it actually went down very slightly, -10kb, maybe because of the
equality comparisons I deleted), so I don't expect it to introduce any
overhead. The output files look the same.

Bug: todo
Change-Id: Ie41d1bf6f6e3197566110774cfb5f20d00d85682