UPSTREAM: Decouple cpu_ipc_ratio and cpu_capacity
Currently, cpu_capacity is both passed to the guest kernel as a hint
about the sizes of the vCPUs relative to each other as well as being
used a relative performance hint to the pCPU that its affined to.

However, cpu_capacity can be used to tune for certain scheduler and task
placement behavior that may not reflect accurate per cycle performance
between various CPUs in specific usecases. Decouple the two parameters
and allow users to set them independently.

Bug=b:284367589
Test: tools/presubmit
Change-Id: I8263ecd33224173da6a49aa1b8768a7464517a75
Signed-off-by: David Dai <davidai@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6168527
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>