ANDROID: further shrink rutabaga_gfx patch file
This moves most of the added lines into a `rust_defaults` module, so
that we are less likely to get conflicts.

It also changes the `rust_test` to more closely match the `rust_library`
setup, i.e. it now uses static linking and `gfxstream_unstable`. Even if
it didn't simplify the patch files, this would still be a good idea, we
should be testing the configuration used in production.

Bug: 398839628
Test: TH
Change-Id: I1ffe07e0b44b0a8aaa931c4e60e87db8b520ae2c