fix: Handle optional cmsgs in recvmsg result
The `recvmsg` function now returns a `Result<CmsgIter>`, which can be
`Err` if there are no control messages. This commit updates the code to
correctly handle this optional result by using `r.cmsgs()?.next()`. This
fixes a build break caused by a recent nix package update that changed
the return type of `cmsgs()`.

Change-Id: Idbed66ec3be0af505098a0867b28196e2dde5793